+++ tests/testthat/test-find-redactor.R
@@ -61,7 +61,7 @@ with_mock_api({
     test_that("redact=NULL to override default (and loaded packages)", {
         expect_true("testpkg" %in% names(sessionInfo()$otherPkgs))
         ## Great, but let's kill it when we're done
-        on.exit(detach(package:testpkg))
+        on.exit(detach("package:testpkg", unload=TRUE))
         newmocks2 <- tempfile()
         expect_warning(
             capture_while_mocking(simplify=FALSE, path=newmocks2, redact=NULL, {
@@ -77,5 +77,27 @@ with_mock_api({
                 body=list(username="password"), encode="json")
         })
         expect_equal(content(b), content(a))
+    })
+
+    test_that("Loading a package with pkgload (devtools)", {
+        newmocks3 <- tempfile()
+        expect_false("testpkg" %in% names(sessionInfo()$otherPkgs))
+        on.exit({
+            pkgload::unload("testpkg")
+            unloadNamespace("pkgload")
+        })
+        expect_message(
+            capture_while_mocking(path=newmocks3, {
+                pkgload::load_all("testpkg")
+                expect_true("testpkg" %in% names(sessionInfo()$otherPkgs))
+                r <- GET("http://example.com/get")
+            }),
+            paste0("Using redact.R from ", dQuote("testpkg"))
+        )
+        with_mock_path(newmocks3, {
+            r2 <- GET("http://example.com/get")
+        })
+        ## The resulting mock content is what we injected into it from testpkg
+        expect_identical(content(r2), list(fake=TRUE))
     })
 })